+++ core/src/gui/menus/bandplan.h
@@ -1,6 +1,8 @@
 #pragma once
+#include <module.h>
 
 namespace bandplanmenu {
     void init();
     void draw(void* ctx);
+    ModuleManager::Instance* getInstance();
 };